Hark a shared spinning cube!!!! I added a script on the player to get and remove authority. You won't need to use this directly. I also added a script on the NetworkManager called LocalPlayer that can be used by any object to get and release authority. These are static function so accessible from any object.
I also made the spinning work with a SyncVar but there is a bug so it always fails to start spinning when it has to grab the authority. After it has it, it’s ok. It’s because grabbing the authority takes a bit of time…and I’m assuming it takes no time…